fix types

+++ lib/model/product.dart
@@ -20,7 +20,7 @@ class Product {
     return Product(
         id: json['id'],
         title: json['title'],
-        price: json['price'],
+        price: json['price'].toDouble(),
         description: json['description'],
         category: json['category'],
         image: json['image'],
@@ -57,7 +57,7 @@ class Rating {
 
   factory Rating.fromJson(Map<String, dynamic> json) {
     return Rating(
-      rate: json['rate'],
+      rate: json['rate'].toDouble(),
       count: json['count'],
     );
   }
